* Refactored method reflection tests * Added tests that Class.getDeclaredMethod and Class.getMethod do not return synthetic equals and hashCode methods added by Phosphor * Masked Class.getDeclaredMethod and Class.getMethod to ensure that they do not return synthetic equals and hashCode methods added by Phosphor * Changed "masking" of getMethod to match getDeclaredMethod
