+++ Phosphor/test/edu/columbia/cs/psl/test/phosphor/runtime/MethodReflectionImplicitITCase.java
@@ -1,0 +1,4 @@
+package edu.columbia.cs.psl.test.phosphor.runtime;
+
+public class MethodReflectionImplicitITCase extends MethodReflectionObjTagITCase {
+}
