+++ Phosphor/test/edu/columbia/cs/psl/test/phosphor/runtime/ReflectionObjTagITCase.java
@@ -37,29 +37,6 @@ public class ReflectionObjTagITCase extends BaseMultiTaintClass {
 		float[] fa = new float[4];
 	}
 
-	public static class MethodHolder {
-		// Whether methods should assert that their arguments are tainted
-		boolean checkArgsTainted;
-
-		MethodHolder(boolean checkArgsTainted) {
-			this.checkArgsTainted = checkArgsTainted;
-		}
-
-		public int primitiveParamMethod(boolean taintedBool) {
-			if(checkArgsTainted) {
-				assertNonNullTaint(MultiTainter.getTaint(taintedBool));
-			}
-			return taintedBool ? 2 : 0;
-		}
-
-		public int primitiveArrParamMethod(boolean[] b) {
-			if(checkArgsTainted) {
-				assertNonNullTaint(MultiTainter.getTaint(b[0]));
-			}
-			return b == null ? 0 : b.length;
-		}
-	}
-
 	public static class ConstructorHolder {
 		public boolean bool;
 		public boolean[] bools;
@@ -108,51 +85,6 @@ public class ReflectionObjTagITCase extends BaseMultiTaintClass {
 			assertTrue(objTaint.contains(valTaint));
 			assertTrue(valTaint.contains(objTaint));
 		}
-	}
-
-	@Test
-	public void testInvokeMethodPrimitiveArg() throws Exception {
-		MethodHolder holder = new MethodHolder(false);
-		Method method = MethodHolder.class.getMethod("primitiveParamMethod", Boolean.TYPE);
-		boolean z = true;
-		Object result = method.invoke(holder, z);
-		assertTrue("Expected integer return from reflected method.", result instanceof Integer);
-		int i = (Integer)result;
-		assertEquals(2, i);
-	}
-
-	@Test
-	public void testInvokeMethodTaintedPrimitiveArg() throws Exception {
-		MethodHolder holder = new MethodHolder(true);
-		Method method = MethodHolder.class.getMethod("primitiveParamMethod", Boolean.TYPE);
-		boolean z = MultiTainter.taintedBoolean(true, "PrimArgLabel");
-		Object result = method.invoke(holder, z);
-		assertTrue("Expected integer return from reflected method.", result instanceof Integer);
-		int i = (Integer)result;
-		assertEquals(2, i);
-	}
-
-	@Test
-	public void testInvokeMethodPrimitiveArrArg() throws Exception {
-		MethodHolder holder = new MethodHolder(false);
-		Method method = MethodHolder.class.getMethod("primitiveArrParamMethod", Class.forName("[Z"));
-		boolean[] arr = new boolean[] {true, true};
-		Object result = method.invoke(holder, (Object) arr);
-		assertTrue("Expected integer return from reflected method.", result instanceof Integer);
-		int i = (Integer)result;
-		assertEquals(arr.length, i);
-	}
-
-	@Test
-	public void testInvokeMethodTaintedPrimitiveArrArg() throws Exception {
-		MethodHolder holder = new MethodHolder(true);
-		Method method = MethodHolder.class.getMethod("primitiveArrParamMethod", Class.forName("[Z"));
-		boolean z = MultiTainter.taintedBoolean(true, "PrimArgLabel");
-		boolean[] arr = new boolean[] {z, z};
-		Object result = method.invoke(holder, (Object) arr);
-		assertTrue("Expected integer return from reflected method.", result instanceof Integer);
-		int i = (Integer)result;
-		assertEquals(arr.length, i);
 	}
 
 	@Test
@@ -210,7 +142,7 @@ public class ReflectionObjTagITCase extends BaseMultiTaintClass {
 				expectedNames.size() == actualNames.size() && expectedNames.containsAll(actualNames) && actualNames.containsAll(expectedNames));
 	}
 
-	/* Checks that phosphor parameters are correctly remapped when calling Constructor.getParameterTypes for a Constructor
+	/* Checks that Phosphor parameters are correctly remapped when calling Constructor.getParameterTypes for a Constructor
 	 * with a primitive parameter. */
 	@Test
 	public void testPrimitiveConstructorGetParamTypes() throws Exception {
@@ -218,7 +150,7 @@ public class ReflectionObjTagITCase extends BaseMultiTaintClass {
 		assertArrayEquals(new Class<?>[]{Boolean.TYPE}, cons.getParameterTypes());
 	}
 
-	/* Checks that phosphor parameters are correctly remapped when calling Constructor.getParameterTypes for a Constructor
+	/* Checks that Phosphor parameters are correctly remapped when calling Constructor.getParameterTypes for a Constructor
 	 * with a primitive parameter and a primitive array parameter. */
 	@Test
 	public void testPrimitiveArrayConstructorGetParamTypes() throws Exception {
@@ -226,49 +158,12 @@ public class ReflectionObjTagITCase extends BaseMultiTaintClass {
 		assertArrayEquals(new Class<?>[]{boolean[].class, Boolean.TYPE}, cons.getParameterTypes());
 	}
 
-	/* Checks that phosphor parameters are correctly remapped when calling Constructor.getParameterTypes for a Constructor
+	/* Checks that Phosphor parameters are correctly remapped when calling Constructor.getParameterTypes for a Constructor
 	 * with a primitive parameter, a primitive array parameter, and primitive 2D array parameter. */
 	@Test
 	public void testPrimitive2DArrayConstructorGetParamTypes() throws Exception {
 		Constructor<ConstructorHolder> cons = ConstructorHolder.class.getConstructor(boolean[][].class, boolean[].class, Boolean.TYPE);
 		assertArrayEquals(new Class<?>[]{boolean[][].class, boolean[].class, Boolean.TYPE}, cons.getParameterTypes());
-	}
-
-	/* Checks that phosphor added equals and hashcode methods are hidden from Class.getDeclaredMethods. */
-	@Test
-	public void testHashCodeAndEqualsHiddenFromGetDeclaredMethods() {
-		String[] methodNames = new String[]{"primitiveParamMethod", "primitiveArrParamMethod"};
-		Class<?>[] returnTypes = new Class<?>[]{Integer.TYPE, Integer.TYPE};
-		Class<?>[][] paramTypes = new Class<?>[][]{
-				new Class<?>[]{Boolean.TYPE},
-				new Class<?>[]{boolean[].class},
-		};
-		Method[] methods = MethodHolder.class.getDeclaredMethods();
-		assertEquals(2, methods.length);
-		for(Method method : methods) {
-			boolean methodMatchesExpected = false;
-			for(int i = 0; i < methodNames.length; i++) {
-				if(method.getName().equals(methodNames[i]) && method.getReturnType().equals(returnTypes[i])
-					&& Arrays.equals(method.getParameterTypes(), paramTypes[i])) {
-					methodMatchesExpected = true;
-					break;
-				}
-			}
-			assertTrue(methodMatchesExpected);
-		}
-	}
-
-	/* Checks that phosphor added equals and hashcode methods are replaced by Object.equals and Object.hashCode for
-	 * Class.getMethods. */
-	@Test
-	public void testHashCodeAndEqualsReplacedInGetMethods() throws NoSuchMethodException {
-		HashSet<Method> expected = new HashSet<>();
-		expected.add(MethodHolder.class.getDeclaredMethod("primitiveParamMethod", Boolean.TYPE));
-		expected.add(MethodHolder.class.getDeclaredMethod("primitiveArrParamMethod", boolean[].class));
-		expected.addAll(Arrays.asList(Object.class.getMethods()));
-		Method[] methods = MethodHolder.class.getMethods();
-		HashSet<Method> actual = new HashSet<>(Arrays.asList(methods));
-		assertEquals(expected, actual);
 	}
 
 	/* Checks that parameters passed to getDeclaredConstructor and newInstance for ignored classes like Boolean are not
@@ -286,14 +181,6 @@ public class ReflectionObjTagITCase extends BaseMultiTaintClass {
 	public void testPrimitiveArrayIgnoredConstructor() throws Exception {
 		Constructor<LazyCharArrayObjTags> constructor = LazyCharArrayObjTags.class.getDeclaredConstructor(Taint.class, char[].class);
 		Object result = constructor.newInstance(null, "testPrimitiveArrayIgnoredConstructor".toCharArray());
-		assertNotNull(result);
-	}
-
-	/* Checks that parameters passed to getDeclaredMethod and invoke for ignored classes like Boolean are not remapped. */
-	@Test
-	public void testBooleanIgnoredMethod() throws Exception {
-		Method method = Boolean.class.getDeclaredMethod("toString", boolean.class);
-		String result = (String)method.invoke(null, false);
 		assertNotNull(result);
 	}
 }
